With 'Changed the wait condition in the 'on_init()' function  to wait for the first joint_state message to arrive with ALL the joint states. This is necessary because the kobuki node, implemented in the kobuki base version, publishes the joint states of the wheels before the xs_sdk node is able to publish the arm joint states. If this control is not efficient, a workaround could be to start the kobuki node after a small delay.
Inizialized the variable 'joint_limits' to 0. This allows to control the callback 'joint_state_cb()' so that it doesn't save joint positions before the 'on_init()' function has saved the joint number of the group. This allows a better control on messages with lower number of joint positions than the necessary and prevents errors of 'out_of_range' in the vectors during execution.